# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "KARBONİZGİ",
        "contact": {}
    },
    "host": "https://api.karbonizgi.tr",
    "paths": {
...
        "/detail-score": {
            "get": {
                "description": "Detay skor bilgileri",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Score"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "skor id değeri girilir.",
                        "name": "score_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kullanıcıya ait soru alt başlıklarına göre questiontype:temel soru başlığını ,QuestionSubType : alt soru başlığını ve SubScore ise alt başlığa ait değeri içerir.Bu kısımlar chat kısmı için kullanılır.",
                        "schema": {
                            "$ref": "#/definitions/controllers.DetailResp"
                        }
                    }
                }
            }
        },
...
        "/score": {
            "post": {
                "description": "Karbon Ayak İzi Hesaplanması",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Score"
                ],
                "summary": "User Puan Kısmı",
                "parameters": [
                    {
                        "description": "/company-questions ve /person-questions endpointlerinde belli bir key value değerleri bulunamktadır. Bu değerleri kullanarak question_name ve question_key değerleri eklenir.",
                        "name": "ScoreInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ScoreInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data kısmında kullanıcının karbon ayak izi değeri döner.Chat kısmında bu kısmı kullanacaksın.",
                        "schema": {
                            "$ref": "#/definitions/controllers.ResponseScore"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.DetailResp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.DetailScore"
                    }
                }
            }
        },
        "controllers.DetailScore": {
            "type": "object",
            "properties": {
                "questionSubType": {
                    "type": "string"
                },
                "questionSub_Id": {
                    "type": "integer"
                },
                "questionType": {
                    "type": "string"
                },
                "subScore": {
                    "type": "number"
                }
            }
        },
...
        "controllers.QuestionSub": {
            "type": "object",
            "properties": {
                "score": {
                    "type": "number"
                },
                "sub_key": {
                    "type": "string"
                }
            }
        },
...
        "controllers.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.ResponseScore": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controllers.ScoreResp"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.ScoreInfo": {
            "type": "object",
            "properties": {
                "question_key": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.QuestionSub"
                    }
                },
                "question_name": {
                    "type": "string"
                }
            }
        },
        "controllers.ScoreResp": {
            "type": "object",
            "properties": {
                "score": {
                    "type": "string"
                },
                "score_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}